Modify data type of the functions convert_base

+++ main.h
@@ -39,7 +39,7 @@ int digit_base_16_upper(va_list ptr);
 /*------------------------------------*/
 int restriction_percentage(const char *str);
 int (*match_case(const char *))(va_list);
-int convert_base(int base, int number, int band);
+int convert_base(int base, long int number, int band);
 int print_number(int n);
 
 #endif
